+++ src/crud/update.ts
@@ -19,7 +19,7 @@ export class Update {
    * @param {string} table - The name of the table where the data will be updated.
    * @param {Record<string, any>[]} rows - An array of objects representing the rows to be updated. Each object should contain an 'id' property.
    * @param {string | number} [consumerId] - Optional. The ID of the consumer, updated as `updaterId` column.
-   * @param {string} [consumerName] - Optional. The name of the consumer, updated as `name` column.
+   * @param {string} [consumerName] - Optional. The name of the consumer, updated as `updaterName` column.
    * @returns {{ query: string, args: unknown[] }} An object containing the generated SQL query string and an array of arguments to be used with the query.
    * 
    */
@@ -50,7 +50,7 @@ export class Update {
     for (const p of propsToUse) {
       if (rows[0][p] === undefined) // do not create case if prop is not in the first row
         continue;
-      const colName = p === "consumerId" ? "updaterId" : p === "consumerName" ? "name" : quoteIfUppercase(p);
+      const colName = p === "consumerId" ? '"updaterId"' : p === "consumerName" ? '"updaterName"' : quoteIfUppercase(p);
       query += `${colName} = CASE `;
       for (let j = 0; j < l; j++) {
         const row = rows[j];